Re-introduce improved command wrapper
Python setuptools is inclued in org.freedesktop.Sdk but not
org.freedesktop.Platform (the runtime). Unfortunately this was not
detected earlier because testing was misstakenly limited to running
with --devel, which mounts a different delevopment runtime.

That left me two choices, include setuptools and appdirs (even more
stuff bundled) or improve upon the wrapper.

This was in part inspired by the PulseEffects project:
https://github.com/wwmm/pulseeffects/blob/master/pulseeffects