Fix various blocking issues with CHIRP
These include patching the application and setup.py file:

- Support XDG Base Directory Specification
  (in favor of adding --persist=~/.chirp)
- Support /app prefix
- Install locales in the correct location, so that locale extensions is
  made possible
- Install the scalable SVG version of the appliction icon, instead of
  the inflexible single PNG file
- Install icons to the correct locations

Due to the SVG icon not being included in the upstream tarball, it
currently has to be added separately. The plan is to upstream these
changes and inform the developers that the SVG should be included in the
tarball for packaging purposes.